+++ pkg/server/database_disabled.go
@@ -51,7 +51,7 @@ func replayByID(id int) ([]byte, error) {
 	return nil, nil
 }
 
-func recordGameResult(g *bgammon.Game, winType int, account1 int, account2 int, replay [][]byte) error {
+func recordGameResult(g *bgammon.Game, winType int8, account1 int, account2 int, replay [][]byte) error {
 	return nil
 }
